+++ office365/sharepoint/fields/builtin_field_id.py
@@ -1,38 +1,49 @@
+"""GUIDs of the built-in SharePoint fields (MS-WSSTS "Data Fields").
+
+Complements :class:`~office365.sharepoint.fields.builtin_field_name.SPBuiltInFieldName`
+(names) with the stable field IDs from the MS-WSSTS specification
+(https://learn.microsoft.com/en-us/openspecs/sharepoint_protocols/ms-wssts/83747ceb-f9a5-456b-bc45-692671c6ce6b).
+"""
+
+from __future__ import annotations
+
+
 class SPBuiltInFieldId:
-    """Retrieves identifiers for the fields that ship with Microsoft SharePoint Foundation."""
+    """GUIDs of the fields that ship with Microsoft SharePoint Foundation."""
 
-    Author = "Author"
-    """Identifies a field that contains the specified author of the SharePoint Foundation object."""
-
-    Created = "Created"
-    """Identifies a field that contains the date and time when the specified SharePoint Foundation object was created"""
-
-    FSObjType = "FSObjType"
-    """Identifies a field that contains information about the file system type"""
-
-    FileDirRef = "FileDirRef"
-    """Identifies a field that contains information about the file directory for the specified
-    SharePoint Foundation object."""
-
-    FileLeafRef = "FileLeafRef"
-    """Identifies a field that contains information about the server-relative URL for the file node that is
-    associated with the specified SharePoint Foundation object."""
-
-    FileType = "FileType"
-    """Identifies a field that contains information about the file type for version history of the specified
-    SharePoint Foundation library picture object."""
-
-    FirstName = "FirstName"
-    """Identifies a field that contains the first name for the specified SharePoint Foundation contact object."""
-
-    FormData = "FormData"
-    """Identifies a field that contains the FormData property for the specified
-    Microsoft SharePoint Foundation contact object."""
-
-    FormURN = "FormURN"
-    """Identifies a field that contains the FormURN property for the specified
-    Microsoft SharePoint Foundation contact object."""
-
-    fRecurrence = "fRecurrence"
-    """Identifies a field that contains information about the recurrence field of the specified SharePoint Foundation
-    calendar event object."""
+    _CopySource = "{6b4e226d-3d88-4a36-808d-a129bf52bccf}"
+    _HasCopyDestinations = "{26d0756c-986a-48a7-af35-bf18ab85ff4a}"
+    _IsCurrentVersion = "{c101c3e7-122d-4d4d-bc34-58e94a38c816}"
+    _Level = "{43bdd51b-3c5b-4e78-90a8-fb2087f71e70}"
+    _ModerationStatus = "{fdc3b2ed-5bf2-4835-a4bc-b885f3396a61}"
+    _UIVersion = "{7841bf41-43d0-4434-9f50-a673baef7631}"
+    _UIVersionString = "{dce8262a-3ae9-45aa-aab4-83bd75fb738a}"
+    Author = "{1df5e554-ec7e-46a6-901d-d85a3881cb18}"
+    BaseName = "{7615464b-559e-4302-b8e2-8f440b913101}"
+    ContentType = "{c042a256-787d-4a6f-8a8a-cf6ab767f12d}"
+    ContentTypeId = "{03e45e84-1992-4d42-9116-26f756012634}"
+    Created = "{8c06beca-0777-48f7-91c7-6da68bc07b69}"
+    Created_x0020_Date = "{998b5cff-4a35-47a7-92f3-3914aa6aa4a2}"
+    Editor = "{d31655d1-1d5b-4511-95a1-7a09e9b75bf2}"
+    EncodedAbsUrl = "{7177cfc7-f399-4d4d-905d-37dd51bc90bf}"
+    File_x0020_Type = "{39360f11-34cf-4356-9945-25c44e68dade}"
+    FileDirRef = "{56605df6-8fa1-47e4-a04c-5b384d59609f}"
+    FileLeafRef = "{8553196d-ec8d-4564-9861-3dbe931050c8}"
+    FileRef = "{94f89715-e097-4e8b-ba79-ea02aa8b7adb}"
+    FolderChildCount = "{960ff01f-2b6d-4f1b-9c3f-e19ad8927341}"
+    FSObjType = "{30bb605f-5bae-48fe-b4e3-1f81d9772af9}"
+    ID = "{1d22ea11-1e32-424e-89ab-9fedbadb6ce1}"
+    InstanceID = "{50a54da4-1528-4e67-954a-e2d24f1e9efb}"
+    ItemChildCount = "{b824e17e-a1b3-426e-aecf-f0184d900485}"
+    Last_x0020_Modified = "{173f76c8-aebd-446a-9bc9-769a2bd2c18f}"
+    MetaInfo = "{687c7f94-686a-42d3-9b67-2782eac4b4f8}"
+    Modified = "{28cf69c5-fa48-462a-b5cd-27b6f9d2bd5f}"
+    Order = "{ca4addac-796f-4b23-b093-d2a3f65c0774}"
+    owshiddenversion = "{d4e44a66-ee3a-4d02-88c9-4ec5ff3f4cd5}"
+    PermMask = "{BA3C27EE-4791-4867-8821-FF99000BAC98}"
+    ProgId = "{c5c4b81c-f1d9-4b43-a6a2-090df32ebb68}"
+    ScopeId = "{dddd2420-b270-4735-93b5-92b713d0944d}"
+    ServerUrl = "{105f76ce-724a-4bba-aece-f81f2fce58f5}"
+    SortBehavior = "{423874f8-c300-4bfb-b7a1-42e2159e3b19}"
+    SyncClientId = "{6d2c4fde-3605-428e-a236-ce5f3dc2b4d4}"
+    UniqueId = "{4b7403de-8d94-43e8-9f0f-137a3e298126}"